jwt/requirements.txt
flask==2.0.1
Flask-JWT-Extended==4.2.3
python-dotenv==0.19.0
